refactor: extract utils from etna.analysis.feature_relevance.plots
